Regen bindings, update notes, and fix a small maths bug

+++ src/Vulkan/Silk.NET.Vulkan/Enums/PhysicalDeviceSchedulingControlsFlagsARM.gen.cs
@@ -11,7 +11,7 @@ namespace Silk.NET.Vulkan
 {
     [Flags]
     [NativeName("Name", "VkPhysicalDeviceSchedulingControlsFlagsARM")]
-    public enum PhysicalDeviceSchedulingControlsFlagsARM : int
+    public enum PhysicalDeviceSchedulingControlsFlagsARM : long
     {
         [NativeName("Name", "")]
         None = 0,
